Parallel DFS soln added/working - next steps: input from user for dedicated input file/algorithm
