+++ src/sudoku_solver.h
@@ -6,24 +6,25 @@
 #include <vector>
 #include <thread>
 #include <omp.h>
+#include <cstring>
+#include <algorithm>
 
 using namespace std;
 
+typedef vector<vector<int>> SudokuGrid;
+
 const int SIZE = 9;
-const int UNASSIGNED = 0;
+
+extern double start;
 
 void readSudokuFromFile(vector<vector<int>> &matrix, const string& filename);
 void printSudoku(vector<vector<int>> &matrix);
 bool isSafeMove(vector<vector<int>> &matrix, int row, int col, int currNum);
 bool solveSudokuSequentialBackTracking(vector<vector<int>> &matrix, int row, int col);
-void parallelBackTrackingSudoku(vector<vector<int>> &matrix, int startNum);
-void parallelBackTrackGateway(vector<vector<int>> &matrix);
 void solveSudoku(vector<vector<int>> &matrix);
-void solveSudokuParallelDFS(vector<vector<int>> &matrix);
-bool parallelDFS(vector<vector<int>> &matrix, int row, int col);
 
-
-bool solve(vector<vector<int>> &grid, int level); 
-void solveSudokuParallelGateway(vector<vector<int>> &matrix);
+bool findZero(const SudokuGrid &grid, int &row, int &col);
+void parallelDFSGateway(SudokuGrid sudoku);
+bool solve(SudokuGrid grid, int level);
 
 #endif // SUDOKU_SOLVER_H
